+++ R/constants.R
@@ -9,6 +9,72 @@
 
 # Define variables to evaluate data consistently across functions
 
+
+# Vector of Harvest Information Program species/species group fields containing
+# bag values
+REF_BAG_FIELDS <-
+  c("ducks_bag", "geese_bag", "dove_bag", "woodcock_bag", "coots_snipe",
+    "rails_gallinules", "cranes", "band_tailed_pigeon", "brant", "seaducks")
+
+REF_STRATA_NAMES <-
+  c("S_ducks", "S_geese", "S_doves", "S_woodcock", "S_coot_snipe",
+    "S_rail_gallinule", "S_cranes", "S_bt_pigeons", "S_brant", "S_seaducks")
+
+# Vector of fields that are used to deduplicate hunters
+REF_HUNTER_ID_FIELDS <-
+  c("firstname", "lastname", "state", "birth_date", "dl_state",
+    "registration_yr")
+
+# US District and Territory abbreviations:
+# District of Columbia, American Samoa, Guam, Northern Mariana Islands,
+# Puerto Rico, Virgin Islands, US Minor Outlying Islands, Marshall Islands,
+# Micronesia, Palau, Armed Forces (Americas), Armed Forces (Europe), Armed
+# Forces (Pacific)
+REF_ABBR_USA <-
+  c("DC", "AS", "GU", "MP", "PR", "VI", "UM", "MH", "FM", "PW", "AA", "AE",
+    "AP")
+
+# Canada abbreviations: Alberta, British Columbia, Manitoba, New Brunswick,
+# Newfoundland and Labrador, Nova Scotia, Northwest Territories, Nunavut,
+# Ontario, Prince Edward Island, Province du Québec, Quebec, Saskatchewan,
+# Yukon; Used by the download report template
+REF_ABBR_CANADA <-
+  c("AB", "BC", "MB", "NB", "NL", "NS", "NT", "NU", "ON", "PE", "PQ", "QC",
+    "SK", "YT")
+
+# Continental 49 state abbreviations; Used by read_hip() and the download report
+REF_ABBR_49_STATES <- datasets::state.abb[datasets::state.abb != "HI"]
+
+# Permit state expected bag values (files received separately from HIP process)
+REF_PMT_FILES <-
+  bind_rows(
+    # Crane permit file states
+    tibble(
+      dl_state = c("CO", "KS", "MN", "MT", "ND", "NM", "OK", "TX", "WY"),
+      spp = "cranes",
+      value = "0"),
+    # Band-tailed Pigeon permit file states
+    tibble(
+      dl_state = c("CO", "NM", "UT"),
+      spp = "band_tailed_pigeon",
+      value = "0")
+  )
+
+# Permit states with in-line permit bag values
+REF_PMT_INLINE <-
+  bind_rows(
+    # Washington in-line permits expected for BTPI, brant, and sea ducks
+    tibble(
+      dl_state = "WA",
+      spp = c("band_tailed_pigeon", "brant", "seaducks"),
+      value = "2"),
+    # Oregon in-line permits expected for BTPI, brant, and sea ducks
+    tibble(
+      dl_state = "OR",
+      spp = c("band_tailed_pigeon", "brant", "seaducks"),
+      value = "2")
+  )
+
 # Define inline permits
 LOGIC_INLINE_PMT <-
   expr(
@@ -76,69 +142,3 @@ REGEX_USA_CANADA <-
   paste(
     c(datasets::state.abb, REF_ABBR_USA, REF_ABBR_CANADA),
     collapse = "|")
-
-# Vector of Harvest Information Program species/species group fields containing
-# bag values
-REF_BAG_FIELDS <-
-  c("ducks_bag", "geese_bag", "dove_bag", "woodcock_bag", "coots_snipe",
-    "rails_gallinules", "cranes", "band_tailed_pigeon", "brant", "seaducks")
-
-REF_STRATA_NAMES <-
-  c("S_ducks", "S_geese", "S_doves", "S_woodcock", "S_coot_snipe",
-    "S_rail_gallinule", "S_cranes", "S_bt_pigeons", "S_brant", "S_seaducks")
-
-# Vector of fields that are used to deduplicate hunters
-REF_HUNTER_ID_FIELDS <-
-  c("firstname", "lastname", "state", "birth_date", "dl_state",
-    "registration_yr")
-
-# US District and Territory abbreviations:
-# District of Columbia, American Samoa, Guam, Northern Mariana Islands,
-# Puerto Rico, Virgin Islands, US Minor Outlying Islands, Marshall Islands,
-# Micronesia, Palau, Armed Forces (Americas), Armed Forces (Europe), Armed
-# Forces (Pacific)
-REF_ABBR_USA <-
-  c("DC", "AS", "GU", "MP", "PR", "VI", "UM", "MH", "FM", "PW", "AA", "AE",
-    "AP")
-
-# Canada abbreviations: Alberta, British Columbia, Manitoba, New Brunswick,
-# Newfoundland and Labrador, Nova Scotia, Northwest Territories, Nunavut,
-# Ontario, Prince Edward Island, Province du Québec, Quebec, Saskatchewan,
-# Yukon; Used by the download report template
-REF_ABBR_CANADA <-
-  c("AB", "BC", "MB", "NB", "NL", "NS", "NT", "NU", "ON", "PE", "PQ", "QC",
-    "SK", "YT")
-
-# Continental 49 state abbreviations; Used by read_hip() and the download report
-REF_ABBR_49_STATES <- datasets::state.abb[datasets::state.abb != "HI"]
-
-# Permit state expected bag values (files received separately from HIP process)
-REF_PMT_FILES <-
-  bind_rows(
-    # Crane permit file states
-    tibble(
-      dl_state = c("CO", "KS", "MN", "MT", "ND", "NM", "OK", "TX", "WY"),
-      spp = "cranes",
-      value = "0"),
-    # Band-tailed Pigeon permit file states
-    tibble(
-      dl_state = c("CO", "NM", "UT"),
-      spp = "band_tailed_pigeon",
-      value = "0")
-  )
-
-# Permit states with in-line permit bag values
-REF_PMT_INLINE <-
-  bind_rows(
-    # Washington in-line permits expected for BTPI, brant, and sea ducks
-    tibble(
-      dl_state = "WA",
-      spp = c("band_tailed_pigeon", "brant", "seaducks"),
-      value = "2"),
-    # Oregon in-line permits expected for BTPI, brant, and sea ducks
-    tibble(
-      dl_state = "OR",
-      spp = c("band_tailed_pigeon", "brant", "seaducks"),
-      value = "2")
-  )
-