+++ R/variables.R
@@ -1,0 +1,4 @@
+# Define variables to evaluate data consistently across functions; these values
+# change with each hunting season
+
+REF_CURRENT_SEASON <- "2024"